Added unit tests for portion of one range or two.

+++ ml/Config.h
@@ -30,7 +30,7 @@ public:
     double ADWindowRateThreshold;
     double ADDimensionRateThreshold;
     //The window size of anomaly bit counting for anomaly percentage
-    double SaveAnomalyPercentageEvery;
+    unsigned long SaveAnomalyPercentageEvery;
 
     SIMPLE_PATTERN *SP_HostsToSkip;
     SIMPLE_PATTERN *SP_ChartsToSkip;
